sizes

+++ 0x00-hello_world/6-size.c
@@ -13,9 +13,9 @@ int main(void)
 	long long allong;
 	float afloat;
 
-	print("Size of a char: %d byte(s)\n", sizeof(achar));
-	printf("Size of an int: %d byte(s)\n", sizeof(aint));
-	printf("Size of a long int: %d byte(s)\n", sizeof(along));
+	print("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(achar));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(aint));
+	printf("Size of a long int: %d byte(s)\n", (unsigned long)sizeof(along));
 	printf("Size of a long long int: %d byte(s)\n", sizeof(allong));
 	printf("Size of a float: %d byte(s)\n", sizeof(afloat));
 	return (0);
